# lib/api-zod/src/generated/types/alarm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlarmAlarmType } from "./alarmAlarmType";
import type { AlarmVibration } from "./alarmVibration";

export interface Alarm {
  id: number;
  name: string;
  enabled: boolean;
  hour: number;
  minute: number;
  alarmType: AlarmAlarmType;
  specificDates?: string[];
  weekDays?: number[];
  autoDelete: boolean;
  ringtone: string;
  volume: number;
  gradualVolume: number;
  vibration: AlarmVibration;
  autoSnooze: number;
  maxSnoozes: number;
  skippedDates?: string[];
  personalTimeKey?: string | null;
  profileId?: number | null;
  iconName?: string;
  nextActivation?: string | null;
  createdAt: string;
  updatedAt: string;
}