# lib/api-zod/src/generated/types/alarmInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlarmInputAlarmType } from "./alarmInputAlarmType";
import type { AlarmInputVibration } from "./alarmInputVibration";

export interface AlarmInput {
  name: string;
  enabled: boolean;
  hour: number;
  minute: number;
  alarmType: AlarmInputAlarmType;
  specificDates?: string[];
  weekDays?: number[];
  autoDelete: boolean;
  ringtone: string;
  volume: number;
  gradualVolume: number;
  vibration: AlarmInputVibration;
  autoSnooze: number;
  maxSnoozes?: number;
  skippedDates?: string[];
  personalTimeKey?: string | null;
  profileId?: number | null;
  iconName?: string;
}