# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileVibration } from "./profileVibration";

export interface Profile {
  id: number;
  name: string;
  isDefault?: boolean;
  autoDelete?: boolean | null;
  ringtone?: string | null;
  volume?: number | null;
  gradualVolume?: number | null;
  vibration?: ProfileVibration;
  autoSnooze?: number | null;
  maxSnoozes?: number | null;
  createdAt: string;
  updatedAt: string;
}