# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsSortOrder } from "./settingsSortOrder";

export interface Settings {
  id: number;
  autoBackup: boolean;
  backupHour: number;
  backupMinute: number;
  backupRetentionDays: number;
  wakeupHour?: number | null;
  wakeupMinute?: number | null;
  workStartHour?: number | null;
  workStartMinute?: number | null;
  lunchHour?: number | null;
  lunchMinute?: number | null;
  workEndHour?: number | null;
  workEndMinute?: number | null;
  dinnerHour?: number | null;
  dinnerMinute?: number | null;
  sleepHour?: number | null;
  sleepMinute?: number | null;
  languageCode?: string | null;
  sortOrder: SettingsSortOrder;
  wakeupLabel?: string | null;
  workStartLabel?: string | null;
  lunchLabel?: string | null;
  workEndLabel?: string | null;
  dinnerLabel?: string | null;
  sleepLabel?: string | null;
  wakeupColor?: string | null;
  workStartColor?: string | null;
  lunchColor?: string | null;
  workEndColor?: string | null;
  dinnerColor?: string | null;
  sleepColor?: string | null;
}